drop fields by list of fields, locally defined
ISTG, this should be a rules engine over json
but serde_json has a recursion limit so we cannot
just recurse down and apply rules as we go.

Maybe this means we need to switch to some other
library that is not dumb, like simd-json.
(it could also mean we write five functions
"recurse_json_lvl_1", "lvl2" ... "lvl5") and call it a day.