fix(sanctioning): correct ordering
if we have a stack of sanctioning passes, they need to be "popped", so
undone in the reverse order they were initially applied, or they will
break each other.

There is still flaws with this placeholder method, where the regex
redaction rules can clobber the placeholder string, but for now this
is probably ok. the symptoms should be obvious in the future if we start
seeing fragmented placeholder strings.